fixed bug of validation

+++ QuizBiz/QuestionChoixRep.h
@@ -28,7 +28,7 @@ public:
 
 	void affichageInstruction() override {
 		std::cout << "\nChoisir une des reponses suivantes en indiquant l'index\n";
-		for (int index{}; index < mChoixReps.size(); ++index)
+		for (size_t index{}; index < mChoixReps.size(); ++index)
 			std::cout << index + 1 << ": " << mChoixReps.at(index) << "\n";
 	}
 private:
